requirements.txt
# Core ML
torch>=2.1.0
torchvision>=0.16.0

# Model zoo
timm>=0.9.0

# GradCAM
grad-cam>=1.5.0

# Image augmentation
albumentations>=1.3.0

# Data / numerics
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0

# Metrics
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0

# NOTE: Install torch with CUDA on server:
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu124
